feat: chamando o service do editar na controller, adicionando o fluxo de excluir o produto.

+++ src/main/java/org/therapazes/luisaoproject/services/ProdutoService.java
@@ -32,6 +32,9 @@ public class ProdutoService {
     }
 
     public void deleteById(Integer id) {
+        if (!produtoRepository.existsById(id)) {
+            throw new EntityNotFoundException("Produto não encontrado");
+        }
         produtoRepository.deleteById(id);
     }
